after chenging ModelAndView to ModelMap

+++ src/main/java/com/student/app/repository/StudentRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.student.app.dto.Student;
 
-public interface StudentRepository extends CrudRepository<Student, Long>{
+public interface StudentRepository extends CrudRepository<Student, Integer>{
 
 	public Student getStudentByEmail(String email);
 }
